refactor(tonic-xds): extract LbPolicy + ring-hash validation into lb_policy
Move the LbPolicy enum and ring_hash_lb_config parsing out of cluster.rs into a
dedicated lb_policy module. The RingHash variant now carries a RingHashSettings
payload whose validate() owns the gRFC A42 rules, so the validation lives with
the config it produces and is unit-testable on its own. cluster.rs re-exports
LbPolicy so existing cluster::LbPolicy paths are unchanged; ClusterResource just
delegates to RingHashSettings::validate.
